Move selection tool in utils from d3-tools
