+++ web/src/components/clustering/TimeSeries.vue
@@ -18,7 +18,7 @@ import { useApiStore, useFileStore } from "@/api/stores";
 import { Cluster } from "@/util/clustering-algorithms/ClusterTypes";
 import { File } from "@/api/models";
 import { useCluster, useD3Tooltip, usePartialLegend } from "@/composables";
-import { SelectionTool, xCoord } from "@/d3-tools/SelectionTool";
+import { SelectionTool, xCoord } from "@/util/SelectionTool";
 import { useElementSize } from "@vueuse/core";
 import * as d3 from "d3";
