+++ web/src/types/imports-components.d.ts
@@ -15,6 +15,10 @@ declare module '@vue/runtime-core' {
     FileTagList: typeof import('./../components/clustering/FileTagList.vue')['default']
     FileTimestamp: typeof import('./../components/FileTimestamp.vue')['default']
     Graph: typeof import('./../components/graph/Graph.vue')['default']
+    GraphElementList: typeof import('./../components/GraphElementList.vue')['default']
+    GraphElementListCard: typeof import('./../components/GraphElementListCard.vue')['default']
+    GraphLegend: typeof import('./../components/GraphLegend.vue')['default']
+    GraphSelectedInfo: typeof import('./../components/GraphSelectedInfo.vue')['default']
     GraphTab: typeof import('./../components/clustering/GraphTab.vue')['default']
     HeatMap: typeof import('./../components/clustering/HeatMap.vue')['default']
     InfoDot: typeof import('./../components/InfoDot.vue')['default']
